# openapi.yaml
# 
# Generated by
# 
#      _____ _          __  __      _     _
#     / ____| |        / _|/ _|    | |   | |
#    | (___ | | ____ _| |_| |_ ___ | | __| | ___ _ __
#     \___ \| |/ / _` |  _|  _/ _ \| |/ _` |/ _ \ '__|
#     ____) |   < (_| | | | || (_) | | (_| |  __/ |
#    |_____/|_|\_\__,_|_| |_| \___/|_|\__,_|\___|_|
#
# The code generator that works in many programming languages
#
#			https://www.skaffolder.com
#
#
# You can generate the code from the command-line
#       https://npmjs.com/package/skaffolder-cli
#
#       npm install -g skaffodler-cli
#
#   #   #   #   #   #   #   #   #   #   #   #   #   #   #   #
#
# To remove this comment please upgrade your plan here: 
#      https://app.skaffolder.com/#!/upgrade
#
# Or get up to 70% discount sharing your unique link:
#       https://app.skaffolder.com/#!/register?friend=5dc9035ef1ef4518a538176b
#
# You will get 10% discount for each one of your friends
# 
#

openapi: 3.0.0
info:
  x-skaffolder-id-project: 5dc903f7f1ef4518a5381771
  title: login
  description: You can obtain JWT token from login API
  version: 0.1
servers:
  - url: http://localhost:3000/api
    description: Local server for testing
paths:
  /login:
    post:
      x-skaffolder-ignore: true
      summary: Login API
      security: []   # No security
      tags:
        - Security
      requestBody:
        content:
          application/json:
            schema:      # Request body contents
              type: object
              properties:
                username:
                  type: string
                password:
                  type: string
              example:   # Sample object
                username: admin
                password: 62f264d7ad826f02a8af714c0a54b197935b717656b80461686d450f7b3abde4c553541515de2052b9af70f710f0cd8a1a2d3f4d60aa72608d71a63a9a93c0f5
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /verifyToken:
    post:
      x-skaffolder-ignore: true
      summary: Verify token API
      security: [] #No security
      tags:
        - Security
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                token:
                  type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /changePassword:
    post:
      x-skaffolder-ignore: true
      summary: Change password API
      tags:
        - Security
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                passwordNew:
                  type: string
                passwordOld:
                  type: string
      responses:
        '200':
          description: OK
                

  #-----  User  -----
  /user/{id}/changePassword:
    post:
      x-skaffolder-id: 5dc903f8f1ef4518a5381780
      x-skaffolder-name: changePassword
      x-skaffolder-id-resource: 5dc903f7f1ef4518a538177b
      x-skaffolder-resource: User
      x-skaffolder-crudAction: 
      x-skaffolder-crudType: 
      x-skaffolder-description: Change password of user from admin
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: object
      x-skaffolder-url: /{id}/changePassword
      x-skaffolder-roles: ["ADMIN"]
      summary: Change password of user from admin
      tags:
        - User
      responses:
        "200":
          description: OK
    
  /user/:
    post:
      x-skaffolder-id: 5dc903f8f1ef4518a538177e
      x-skaffolder-name: create
      x-skaffolder-id-resource: 5dc903f7f1ef4518a538177b
      x-skaffolder-resource: User
      x-skaffolder-crudAction: create
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION create
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /
      x-skaffolder-roles: []
      summary: CRUD ACTION create
      tags:
        - User
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                mail:
                  type: string
                name:
                  type: string
                password:
                  type: string
                roles:
                  type: string
                surname:
                  type: string
                username:
                  type: string
    get:
      x-skaffolder-id: 5dc903f8f1ef4518a538177f
      x-skaffolder-name: list
      x-skaffolder-id-resource: 5dc903f7f1ef4518a538177b
      x-skaffolder-resource: User
      x-skaffolder-crudAction: list
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION list
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /
      x-skaffolder-roles: []
      summary: CRUD ACTION list
      tags:
        - User
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
    
  /user/{id}:
    delete:
      x-skaffolder-id: 5dc903f8f1ef4518a5381781
      x-skaffolder-name: delete
      x-skaffolder-id-resource: 5dc903f7f1ef4518a538177b
      x-skaffolder-resource: User
      x-skaffolder-crudAction: delete
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION delete
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION delete
      tags:
        - User
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
    get:
      x-skaffolder-id: 5dc903f8f1ef4518a5381785
      x-skaffolder-name: get
      x-skaffolder-id-resource: 5dc903f7f1ef4518a538177b
      x-skaffolder-resource: User
      x-skaffolder-crudAction: get
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION get
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION get
      tags:
        - User
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id resource
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    post:
      x-skaffolder-id: 5dc903f8f1ef4518a5381783
      x-skaffolder-name: update
      x-skaffolder-id-resource: 5dc903f7f1ef4518a538177b
      x-skaffolder-resource: User
      x-skaffolder-crudAction: update
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION update
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION update
      tags:
        - User
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/User"
    
  #-----  user  -----
  /user/:
    post:
      x-skaffolder-id: 5dc90540f1ef4518a53817d6
      x-skaffolder-name: create
      x-skaffolder-id-resource: 5dc90530f1ef4518a53817d1
      x-skaffolder-resource: user
      x-skaffolder-crudAction: create
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION create
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /
      x-skaffolder-roles: []
      summary: CRUD ACTION create
      tags:
        - user
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/user"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                123:
                  type: number
    get:
      x-skaffolder-id: 5dc90541f1ef4518a53817da
      x-skaffolder-name: list
      x-skaffolder-id-resource: 5dc90530f1ef4518a53817d1
      x-skaffolder-resource: user
      x-skaffolder-crudAction: list
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION list
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /
      x-skaffolder-roles: []
      summary: CRUD ACTION list
      tags:
        - user
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/user"
    
  /user/{id}:
    delete:
      x-skaffolder-id: 5dc90541f1ef4518a53817db
      x-skaffolder-name: delete
      x-skaffolder-id-resource: 5dc90530f1ef4518a53817d1
      x-skaffolder-resource: user
      x-skaffolder-crudAction: delete
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION delete
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION delete
      tags:
        - user
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
    get:
      x-skaffolder-id: 5dc90540f1ef4518a53817d7
      x-skaffolder-name: get
      x-skaffolder-id-resource: 5dc90530f1ef4518a53817d1
      x-skaffolder-resource: user
      x-skaffolder-crudAction: get
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION get
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION get
      tags:
        - user
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id resource
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/user"
    post:
      x-skaffolder-id: 5dc90540f1ef4518a53817d4
      x-skaffolder-name: update
      x-skaffolder-id-resource: 5dc90530f1ef4518a53817d1
      x-skaffolder-resource: user
      x-skaffolder-crudAction: update
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION update
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION update
      tags:
        - user
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/user"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/user"
    
    
# JWT security
security:
  - bearerAuth: [] 
  
  
# Components
components:

  # Security
  securitySchemes:
    bearerAuth:            # arbitrary name for the security scheme
      type: http
      scheme: bearer
      bearerFormat: JWT    # optional, arbitrary value for documentation purposes

  # Data Model
  schemas:
        

    User:
      x-skaffolder-id: 5dc903f7f1ef4518a538177b
      x-skaffolder-id-db: 5dc903f7f1ef4518a5381772
      x-skaffolder-id-entity: 5dc903f7f1ef4518a5381773
      x-skaffolder-url: /user
      properties:
        _id:
          type: string
          x-skaffolder-required: true
        mail:
          type: string
          x-skaffolder-id-attr: 5dc903f7f1ef4518a5381776
          x-skaffolder-type: String
        name:
          type: string
          x-skaffolder-id-attr: 5dc903f7f1ef4518a5381777
          x-skaffolder-type: String
        password:
          type: string
          x-skaffolder-id-attr: 5dc903f7f1ef4518a5381775
          x-skaffolder-type: String
          x-skaffolder-required: true
        roles:
          type: string
          x-skaffolder-id-attr: 5dc903f7f1ef4518a5381779
          x-skaffolder-type: String
        surname:
          type: string
          x-skaffolder-id-attr: 5dc903f7f1ef4518a5381778
          x-skaffolder-type: String
        username:
          type: string
          x-skaffolder-id-attr: 5dc903f7f1ef4518a5381774
          x-skaffolder-type: String
          x-skaffolder-required: true
      x-skaffolder-relations: 
      required:
        - _id
        - password
        - username
      
    
    user:
      x-skaffolder-id: 5dc90530f1ef4518a53817d1
      x-skaffolder-id-db: 5dc903f7f1ef4518a5381772
      x-skaffolder-id-entity: 5dc90530f1ef4518a53817ce
      x-skaffolder-url: /user
      properties:
        _id:
          type: string
          x-skaffolder-required: true
        123:
          type: number
          x-skaffolder-id-attr: 5dc90530f1ef4518a53817cf
          x-skaffolder-type: Number
      x-skaffolder-relations: 
        log:
          x-skaffolder-type: m:m
          x-skaffolder-ent1: 5dc90530f1ef4518a53817ce
          x-skaffolder-ent2: 5dc903f7f1ef4518a5381773
      required:
        - _id
      
    
    
  # Databases
  x-skaffolder-db:
    - x-skaffolder-id: 5dc903f7f1ef4518a5381772
      x-skaffolder-name: login_db

  # Pages
  x-skaffolder-page:
    - x-skaffolder-id: 5dc903f7f1ef4518a538177a
      x-skaffolder-name: Home
      x-skaffolder-url: /home
      x-skaffolder-template: 
      x-skaffolder-resource: 
      x-skaffolder-services: 
      x-skaffolder-nesteds: 
      x-skaffolder-links: 
        - 5dc90541f1ef4518a53817d9
      x-skaffolder-roles: []
    - x-skaffolder-id: 5dc90540f1ef4518a53817d3
      x-skaffolder-name: userEdit
      x-skaffolder-url: /users/{id}
      x-skaffolder-template: Edit_Crud
      x-skaffolder-resource: 5dc90530f1ef4518a53817d1
      x-skaffolder-services: 
        - 5dc90540f1ef4518a53817d6
        - 5dc90540f1ef4518a53817d7
        - 5dc903f8f1ef4518a538177f
        - 5dc90540f1ef4518a53817d4
      x-skaffolder-nesteds: 
      x-skaffolder-links: 
      x-skaffolder-roles: []
    - x-skaffolder-id: 5dc90541f1ef4518a53817d9
      x-skaffolder-name: userList
      x-skaffolder-url: /users
      x-skaffolder-template: List_Crud
      x-skaffolder-resource: 5dc90530f1ef4518a53817d1
      x-skaffolder-services: 
        - 5dc90541f1ef4518a53817db
        - 5dc90541f1ef4518a53817da
      x-skaffolder-nesteds: 
      x-skaffolder-links: 
        - 5dc90540f1ef4518a53817d3
      x-skaffolder-roles: []